Expandable list in preperation activity
git-svn-id: https://faap-feedme.googlecode.com/svn/trunk@31 c03250ec-a590-73f7-ba79-85b89e17e35f

+++ FeedMe/src/de/faap/feedme/util/Recipe.java
@@ -43,7 +43,7 @@ public class Recipe {
     }
 
     public void changePortions(int newPortions) {
-	double quotient = newPortions / portions;
+	double quotient = (double) newPortions / (double) portions;
 	portions = newPortions;
 	for (int i = 0; i < quantities.length; i++) {
 	    quantities[i] = quantities[i] * quotient;
